avoid equal fragments in test of ReadFromBs
Equal fragments are ordered in random order inside
BlockSet structure (field seq2fragments).
(This occurs in LuaJIT, but not in reference Lua.)
That is why such blocksets are failed to be compared.
Equal fragments are rare and not important case,
so I changed the test, not the code of BlockSet:__eq.