BlockSet: assert against sequence names collision
If a fragment is located on sequence A which name collides
with name of sequence B and A is not in the blockset
and B is in the blockset, then this asser fires.

Previously error on other line of code was fired;
the error is less understandable, than this assert.