cpp, BlockSet: sort by fragment and by block
Previously, if a blockset has equal fragments, it could
be compared not to be equal with copy of itself, because
two pairs (fragment, block) can be ordered arbitrary if
fragments are equal.

Currently, this bug was fixed (by sorting not only by
fragment, but also by block, if fragments are equal).

Test was added.