C++, BlockSet: use segment tree to find overlaps
Simple sorted array of fragments is not sufficient if
a fragment is inside other fragment:

    f1      ###########
    f2       #####
    f3        ###########
    pattern        ????

If this is detected (if order of max positions disagrees
with order of min positions), then additional data structure
is created: segment tree [1].

Non-leaf nodes of segment tree are stored in array of size
N-1, where N is number of fragments.

New test was added: many random overlapping fragments.

[1] https://en.wikipedia.org/wiki/Segment_tree