consensus, unwind: use block names instead of map
ConsensusSequences used to return map seqs2blocks. This map
was used by block.unwind and algo.UnwindBlocks. Currently
ConsensusSequences doesn't return this map. It assigns blocks
names to sequences, so this map is not needed. To distinguish
blocks with same names from different blocksets, a blockset
gets a prefix. Default prefix is "" and is used for one
blockset case. Sequence_name = prefix + block_name. A map from
prefix to blockset is passed to block.unwind and
algo.UnwindBlocks.