ShortForm.decode: do not keep all fragment texts
Lua's memory is limited (1 Gb for LuaJIT). Previous version of
ShortForm kept all texts in memory at once. Currently it keeps
only consensuses + diff instead of text. Drawback is that texts
are calculated twice: for sequences and for blocks.

Probably better solution would be moving frid2text map to C++.
Memory of C++ objects is not included in Lua's memory.