change alignment.refine to pass pending tests
A base used not to be moved if 'from' column has only that
base and gap. For example:

A-
-A
AA

Nothing was moved according to this rule.

Currently it was adjusted: do not move a base if 'from' column
has only that base and gap AND number of bases of that sort in
'from' column if greater than that number in 'to' column.