considers good block of short fragments to be good
If block consists of fragments shorter than the theshold, but
the length of a block is more or equal to the treshold and an
alignment (local identity + ends) of a block is good, it must
be considered good, not minor.

It is stupid to make it a minor block, because it is already
considered good by isGood and goodSubblocks. Alternatively
we have to change isGood and goodSubblocks to consider lengths
of fragments instead of lengths of blocks, which we have refused
to do because it doesn't change the essence of the result, but
requires too complex algorithms (e.g., goodSubblocks).